There is a bit too much here refactor/finish some moves
There is a bit too much here for one commit. This is a result of me
starting a feature the last time I was working and then cutting it short
and pushing to save my work. Here I just piked up where I left off and
in the heat of the moment keep working to refactor some of this code.
Hopefully the next commit will put me back on pace.